flashcards/requirements.txt
Django>=4.2,<5.0
PyPDF2>=3.0.0
python-docx>=1.0.0
stripe>=7.0.0

